support 64bit in ndk components per Google Play requirements fix issues #36, #39, #41, #42, #45

+++ app/src/main/java/org/andglkmod/glk/BlankWindow.java
@@ -68,7 +68,7 @@ public class BlankWindow extends Window {
 	}
 
 	@Override
-	public void requestLineEvent(String initial, long maxlen, int buffer, int unicode) {
+	public void requestLineEvent(String initial, long maxlen, long buffer, int unicode) {
 	}
 
 	@Override
